+++ CHANGELOG.md
@@ -19,6 +19,17 @@ __________ DO NOT TOUCH ___________
 
 __________ DO NOT TOUCH ___________
 
+s
+## [22.12.0]
+
+### Added
+- Add new model for qc_metrics
+- Add new model for mip_analysis
+
+### Changed
+- Use new metrics deliverables format, which is already part of qc_metrics file
+- Use new models when parsing mip analysis files
+
 ## [22.11.3]
 ### Fixed
 - Show more errors from parsing orderforms